Update Footer component with new design and improved styling consistency
